+++ Code/Engine/Foundation/Math/Angle.h
@@ -73,6 +73,9 @@ public:
   bool operator == (const ezAngle& r) const; // [tested]
   bool operator != (const ezAngle& r) const; // [tested]
 
+  // At least the < operator is implement to make clamping etc. work
+  bool operator< (const ezAngle& r) const;
+
   // Note: relational operators on angles are not really possible - is 0 degree smaller or bigger than 359 degree?
 
 private: